Rename Fixtures.hpp to LoggerFixtures.hpp
