+++ tests/unit/data/AmendmentCenterTests.cpp
@@ -19,7 +19,7 @@
 
 #include "data/AmendmentCenter.hpp"
 #include "data/Types.hpp"
-#include "util/Fixtures.hpp"
+#include "util/LoggerFixtures.hpp"
 #include "util/MockBackendTestFixture.hpp"
 #include "util/MockPrometheus.hpp"
 #include "util/TestObject.hpp"
