+++ tests/unit/etl/ETLStateTests.cpp
@@ -18,7 +18,7 @@
 //==============================================================================
 
 #include "etl/ETLState.hpp"
-#include "util/Fixtures.hpp"
+#include "util/LoggerFixtures.hpp"
 #include "util/MockSource.hpp"
 
 #include <boost/json/parse.hpp>
