+++ tests/unit/etl/GrpcSourceTests.cpp
@@ -18,7 +18,7 @@
 //==============================================================================
 
 #include "etl/impl/GrpcSource.hpp"
-#include "util/Fixtures.hpp"
+#include "util/LoggerFixtures.hpp"
 #include "util/MockBackend.hpp"
 #include "util/MockPrometheus.hpp"
 #include "util/MockXrpLedgerAPIService.hpp"
