+++ tests/unit/rpc/BaseTests.cpp
@@ -24,7 +24,7 @@
 #include "rpc/common/Types.hpp"
 #include "rpc/common/ValidationHelpers.hpp"
 #include "rpc/common/Validators.hpp"
-#include "util/Fixtures.hpp"
+#include "util/LoggerFixtures.hpp"
 
 #include <boost/json/array.hpp>
 #include <boost/json/object.hpp>
