+++ tests/unit/rpc/CountersTests.cpp
@@ -20,7 +20,7 @@
 #include "rpc/Counters.hpp"
 #include "rpc/JS.hpp"
 #include "rpc/WorkQueue.hpp"
-#include "util/Fixtures.hpp"
+#include "util/LoggerFixtures.hpp"
 #include "util/MockPrometheus.hpp"
 #include "util/prometheus/Counter.hpp"
